+++ src/FreeElephants/PsrRouter/Exception/NotFound.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace FreeElephants\PsrRouter\Exception;
+
+class NotFound extends \RuntimeException implements ExceptionInterface
+{
+
+}