+++ src/FreeElephants/PsrRouter/MethodNotAllowedHandlerInterface.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace FreeElephants\PsrRouter;
+
+use Psr\Http\Server\RequestHandlerInterface;
+
+interface MethodNotAllowedHandlerInterface extends RequestHandlerInterface
+{
+
+}